+++ examples/domain-verification/main.tf
@@ -1,10 +1,17 @@
 module "ses" {
-  source   = "../.."
+  source = "../.."
+
+  prefix      = var.prefix
+  environment = var.environment
+  tags        = var.tags
+
   ses_mode = "domain"
   ses_domain = {
-    domain            = "example.com"
+    domain            = "mail.domain.com"
     is_verify_dkim    = true
     is_verify_domain  = true
-    route53_zone_name = "example.com"
+    route53_zone_name = "domain.com"
   }
+
+  is_create_consumer_policy = true
 }