+++ variables.tf
@@ -1,3 +1,27 @@
+/* -------------------------------------------------------------------------- */
+/*                                   Generic                                  */
+/* -------------------------------------------------------------------------- */
+
+variable "prefix" {
+  description = "The prefix name of customer to be displayed in AWS console and resource"
+  type        = string
+}
+
+variable "environment" {
+  description = "To manage a resources with tags"
+  type        = string
+}
+
+variable "tags" {
+  description = "Custom tags which can be passed on to the AWS resources. They should be key value pairs having distinct keys."
+  type        = map(string)
+  default     = {}
+}
+
+/* -------------------------------------------------------------------------- */
+/*                                     SES                                    */
+/* -------------------------------------------------------------------------- */
+
 variable "ses_mode" {
   description = "Mode defines which method to verify identity for SES, which are email and domain."
   type        = string
@@ -32,4 +56,10 @@ variable "ses_domain" {
     is_verify_domain  = false
     route53_zone_name = null
   }
+}
+
+variable "is_create_consumer_policy" {
+  description = "Whether to create consumer readonly policy"
+  type        = bool
+  default     = false
 }