+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this module will be documented in this file.
 
+## [1.0.1] - 2022-10-04
+
+### Added
+
+- `is_create_consumer_policy` to enabled create the policy to grant permission for sending email
+- output `cosumer_policy_arn`
+
 ## [1.0.0] - 2022-07-22
 
 ### Added